@@ -1,6 +1,7 @@
 {
   "name": "miguelquispe.com",
-  "version": "0.1.0",
+  "description": "Miguel Quispe's personal website",
+  "version": "2.0.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -9,33 +10,37 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@types/glob": "^8.0.0",
-    "@vercel/analytics": "^0.1.8",
-    "clsx": "^1.2.1",
-    "date-fns": "^2.29.2",
-    "glob": "^8.0.3",
+    "@tailwindcss/postcss": "^4.0.14",
+    "@types/glob": "^8.1.0",
+    "@vercel/analytics": "^1.5.0",
+    "clsx": "^2.1.1",
+    "date-fns": "^4.1.0",
+    "glob": "^11.0.1",
     "gray-matter": "^4.0.3",
-    "next": "12",
-    "next-mdx-remote": "^4.1.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-share": "^4.4.1",
+    "next": "^15.2.4",
+    "next-mdx-remote": "^5.0.0",
+    "postcss": "^8.5.3",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-share": "^5.2.2",
     "reading-time": "^1.5.0",
-    "rehype-autolink-headings": "^6.1.1",
-    "rehype-code-titles": "1.1.0",
-    "rehype-prism-plus": "^1.5.0",
-    "rehype-slug": "^5.0.1",
-    "remark-gfm": "^3.0.1",
-    "sass": "^1.32.8"
+    "rehype-autolink-headings": "^7.1.0",
+    "rehype-code-titles": "^1.2.0",
+    "rehype-prism-plus": "^2.0.0",
+    "rehype-slug": "^6.0.0",
+    "remark-gfm": "^4.0.1",
+    "sass": "^1.86.0",
+    "tailwindcss": "^4.0.14"
   },
   "devDependencies": {
-    "@types/node": "^18.7.5",
-    "@types/react": "^18.0.17",
-    "eslint": "8.21.0",
-    "eslint-config-next": "12.2.5",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "prettier": "^2.8.4",
-    "typescript": "^4.7.4"
+    "@next/eslint-plugin-next": "^15.3.2",
+    "@types/node": "^22.13.10",
+    "@types/react": "^19.1.0",
+    "eslint": "^9.22.0",
+    "eslint-config-next": "^15.2.3",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-plugin-prettier": "^5.2.3",
+    "prettier": "^3.5.3",
+    "typescript": "^5.8.2"
   }
 }
