@@ -1,7 +1,4 @@
 import clsx from 'clsx';
-import React from 'react';
-
-import styles from './Container.module.scss';
 
 interface Props {
   children: React.ReactNode;
@@ -9,7 +6,7 @@ interface Props {
 }
 
 const Container = ({ children, small }: Props) => {
-  let classNames = clsx(styles.container, small && styles.small);
+  const classNames = clsx('mx-auto', small ? 'max-w-4xl' : 'max-w-5xl');
 
   return <div className={classNames}>{children}</div>;
 };
