@@ -1,16 +1,17 @@
 import Link from 'next/link';
+import { AnchorHTMLAttributes } from 'react';
 
-const PostLink = (props: any) => {
+type PostLinkProps = {
+  href: string;
+} & Omit<AnchorHTMLAttributes<HTMLAnchorElement>, 'href'>;
+
+const PostLink = (props: PostLinkProps) => {
   const href = props.href;
   const isInternalLink = href && href.startsWith('/');
   const isAnchorLink = href && href.startsWith('#');
 
   if (isInternalLink) {
-    return (
-      <Link href={href} {...props}>
-        {props.children}
-      </Link>
-    );
+    return <Link {...props}>{props.children}</Link>;
   }
 
   if (isAnchorLink) {
