@@ -14,12 +14,12 @@ export const PROJECTS: Project[] = [
     image: '/images/portfolio/frontend-canastas.png',
     alt: 'Proyecto web E-commerce',
     description: 'Desarrollo Frontend para web E-commerce.',
-    tags: ['Frontend', 'E-commerce'],
+    tags: ['WordPress', 'Frontend', 'E-commerce'],
   },
   {
     title: 'The Happy Beauty',
     image: '/images/portfolio/frontend-thb.png',
-    alt: 'Proyecto frontend web app',
+    alt: 'Proyecto React sNext.js web app',
     description: 'Desarrollo Frontend para aplicación web.',
     tags: ['Frontend', 'Next.js', 'TypeScript', 'Web App'],
   },
